Fixed displays product qtty in differents users

+++ src/main/java/ua/lviv/lgs/dto/BucketDto.java
@@ -3,13 +3,13 @@ package ua.lviv.lgs.dto;
 public class BucketDto {
 
 	public String bucketId;
-	public Integer productId;
+	public Long productId;
 	public byte[] bs;
 	public String productName;
 	public Double productPrice;
 	public Integer productsCount;
 
-	public BucketDto(String bucketId, byte[] bs, Integer productId, String productName, Double productPrice,
+	public BucketDto(String bucketId, byte[] bs, Long productId, String productName, Double productPrice,
 			Integer productsCount) {
 		this.bucketId = bucketId;
 		this.bs = bs;
